✨ 채팅방 강퇴버튼 조건부 출력(퇴장기능x)
