+++ src/components/chat/ChatToggle.jsx
@@ -1,12 +1,27 @@
+import { useEffect, useRef } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import instance from "../../assets/constants/instance";
 
 const ChatToggle = (props) => {
-  const toggleRef = props.toggleRef;
+  const memberRef = useRef(null);
+  const toggleRef = useRef(null);
   const setToggle = props.setToggle;
   const members = props.members;
   const { concertTitle, id } = useParams();
   const navigate = useNavigate();
+
+  //토글 동작 함수
+  useEffect(() => {
+    const toggleClose = (e) => {
+      if (toggleRef.current && !toggleRef.current.contains(e.target)) {
+        setToggle(false);
+      }
+    };
+    document.addEventListener("click", toggleClose);
+    return () => {
+      document.removeEventListener("click", toggleClose);
+    };
+  }, []);
 
   const exitChatroom = async () => {
     try {
@@ -22,14 +37,32 @@ const ChatToggle = (props) => {
     }
   };
 
-  const chatMembers = [];
+  const forceOut = (e) => {
+    window.alert(e.target.value + "를 강퇴합니다.");
+    return;
+  };
+
+  let chatMembers = [];
   members.map((member) => {
-    chatMembers.push(
+    const isCreator = props.creator === props.me ? "" : "hidden";
+    const $element = (
       <li key={member.username} className="member-info">
-        <div className="member-img">{member.username}</div>
-        <div className="member-name">{member.nickName}</div>
+        <img className="member-img" src={member.profileImage} alt="" />
+        <p className="member-name">{member.nickName}</p>
+        <button
+          className={`force-out ${isCreator}`}
+          onClick={forceOut}
+          value={member.username}
+        >
+          강퇴
+        </button>
       </li>
     );
+    if (member.username === props.me) {
+      chatMembers = [$element, ...chatMembers];
+    } else {
+      chatMembers.push($element);
+    }
   });
   return (
     <div className="toggle-lists" ref={toggleRef}>
@@ -40,7 +73,9 @@ const ChatToggle = (props) => {
         <div className="title" onClick={() => navigate("/")}>
           위드콘
         </div>
-        <ul className="member-lists">{chatMembers}</ul>
+        <ul className="member-lists" ref={memberRef}>
+          {chatMembers}
+        </ul>
       </div>
       <div className="chat-exit">
         <button onClick={exitChatroom}>채팅방 나가기</button>
